Add support for locally installed generator modules.
This commit extends the CLI generator module search functionality
to also look in the current working directory. That way users
of stimpak can install the generator locally before running
the stimpak CLI.

It also refactors the CLI spec setup (creating directories and
symlinks) into a separate helper module to be able to use it in
different specs.